changed key bindings names

+++ Holojam/Assets/Scripts/PlayerController.cs
@@ -124,7 +124,7 @@ public class PlayerController : MonoBehaviour
         //do things like update sprites and playing a character swap animation, ect.
         //change current bullet
         BaseGun gun = GetComponentInChildren<BaseGun>();
-        gun.currentBullet = (BulletType)((int)(currentCharacter) - 1);
+        gun.currentBullet = (BulletType)(currentCharacter - 1);
     }
